UCP/API : Moving ucp am api from ucpx.h
